Add real-world OOXML fixtures from the Open XML SDK and fix bugs they surface
Copy a handful of real, Excel-authored .xlsx files from dotnet/Open-XML-SDK's
own test suite (its "OFCAT"/Robustness corpus, its markup-compatibility
fixture, its extLst fixture and its missing-calcChain fixture) and add
RealWorldFixtureTests, mirroring how the SDK itself uses each file, to
validate them with SchemaValidator and round-trip them through XLWorkbook.

Running the corpus through XlsxSharp surfaced three real save-path bugs,
now fixed:
- A conditional format's differential alignment wrote readingOrder through
  the string-token enum writer instead of casting it to its numeric
  ST_ReadingOrder value, throwing on any dxf alignment that set it.
- A data bar's x14 negativeFillColor/axisColor extension assumed every
  XLColor was RGB, throwing when the negative color was a theme color.
- An autofilter column left at the default "not actively filtered"
  state (FilterType.None) hit an unhandled switch case instead of being
  written bare, which CT_FilterColumn allows.

Also make OpcPackage/OpcPart.GetRelatedParts tolerate a dangling
relationship (target part referenced but never actually written) instead
of throwing, since a real file with e.g. a stale calcChain.xml reference
does turn up and calcChain itself is an optional, Excel-regenerable part.
GetRelatedPart(id) - a direct, by-id lookup - keeps throwing.

+++ XlsxSharp/Excel/IO/StylesWriter.cs
@@ -1078,7 +1078,7 @@ internal class StylesWriter
 
         if (alignment.ReadingOrder is { } readingOrder)
         {
-            xml.WriteAttribute("readingOrder", readingOrder);
+            xml.WriteAttribute("readingOrder", (uint)readingOrder);
         }
 
         xml.WriteEndElement();
